+++ robozonky-app/src/main/java/com/github/robozonky/app/configuration/DaemonOperatingMode.java
@@ -27,11 +27,11 @@ import com.github.robozonky.api.notifications.RoboZonkyInitializedEvent;
 import com.github.robozonky.api.notifications.SessionInfo;
 import com.github.robozonky.app.Events;
 import com.github.robozonky.app.authentication.Authenticated;
+import com.github.robozonky.app.configuration.daemon.BlockedAmountsUpdater;
 import com.github.robozonky.app.configuration.daemon.DaemonInvestmentMode;
+import com.github.robozonky.app.configuration.daemon.PortfolioUpdater;
 import com.github.robozonky.app.investing.Investor;
-import com.github.robozonky.app.portfolio.BlockedAmountsUpdater;
 import com.github.robozonky.app.portfolio.Delinquents;
-import com.github.robozonky.app.portfolio.PortfolioUpdater;
 import com.github.robozonky.common.extensions.MarketplaceLoader;
 import com.github.robozonky.common.secrets.Credentials;
 import com.github.robozonky.common.secrets.SecretProvider;
