+++ robozonky-app/src/main/java/com/github/robozonky/app/configuration/daemon/BlockedAmounts.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.github.robozonky.app.portfolio;
+package com.github.robozonky.app.configuration.daemon;
 
 import java.util.Collection;
 import java.util.Collections;
@@ -25,6 +25,8 @@ import java.util.stream.Collectors;
 
 import com.github.robozonky.api.remote.entities.BlockedAmount;
 import com.github.robozonky.app.authentication.Authenticated;
+import com.github.robozonky.app.portfolio.Portfolio;
+import com.github.robozonky.app.portfolio.PortfolioDependant;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
